rtlwifi: extend debug_comp to u64
The number of debugging conditions now exceeds the capabilities of a
32-bit word.

+++ debug.h
@@ -178,7 +178,7 @@ do {									\
 struct rtl_priv;
 
 __printf(5, 6)
-void _rtl_dbg_trace(struct rtl_priv *rtlpriv, int comp, int level,
+void _rtl_dbg_trace(struct rtl_priv *rtlpriv, u64 comp, int level,
 		    const char *modname, const char *fmt, ...);
 
 #define RT_TRACE(rtlpriv, comp, level, fmt, ...)			\
@@ -217,7 +217,7 @@ static inline void RT_ASSERT(int exp, const char *fmt, ...)
 
 __printf(4, 5)
 static inline void RT_TRACE(struct rtl_priv *rtlpriv,
-			    int comp, int level,
+			    u64 comp, int level,
 			    const char *fmt, ...)
 {
 }
